+++ yeat/tests/test_single.py
@@ -20,7 +20,7 @@ def test_single_end_assemblers_dry_run(tmp_path):
 
 @pytest.mark.long
 @pytest.mark.illumina
-@pytest.mark.parametrize("algorithm", ["spades", "megahit", "unicycler", "velvet"])
+@pytest.mark.parametrize("algorithm", ["spades", "megahit", "unicycler", "penguin", "velvet"])
 def test_single_end_assemblers(algorithm, capsys, tmp_path):
     wd = str(tmp_path)
     config = write_config(algorithm, wd, "single.cfg")